Docs: add documentation links in examples
