+++ examples/envelope.js
@@ -6,6 +6,9 @@
   <button style="margin: 0 2em 2em 0">Play</button>
   Volume: <label>0</label>
   <div id="container" style="border: 1px solid #ddd;"></div>
+  <p>
+    📖 <a href="https://wavesurfer-js.org/docs/classes/plugins_envelope.EnvelopePlugin">Envelope plugin docs</a>
+  </p>
 </html>
 */
