+++ examples/minimap.js
@@ -5,7 +5,7 @@ import Minimap from 'https://unpkg.com/wavesurfer.js@beta/dist/plugins/minimap.j
 
 // Create an instance of WaveSurfer
 const ws = WaveSurfer.create({
-  container: document.body,
+  container: '#waveform',
   waveColor: 'rgb(200, 0, 200)',
   progressColor: 'rgb(100, 0, 100)',
   url: '/examples/audio/audio.wav',
@@ -26,3 +26,12 @@ const ws = WaveSurfer.create({
 ws.on('interaction', () => {
   ws.play()
 })
+
+/*
+<html>
+  <div id="waveform"></div>
+  <p>
+    📖 <a href="https://wavesurfer-js.org/docs/classes/plugins_minimap.MinimapPlugin">Minimap plugin docs</a>
+  </p>
+</html>
+*/
