+++ examples/spectrogram.js
@@ -5,7 +5,7 @@ import Spectrogram from 'https://unpkg.com/wavesurfer.js@beta/dist/plugins/spect
 
 // Create an instance of WaveSurfer
 const ws = WaveSurfer.create({
-  container: document.body,
+  container: '#waveform',
   waveColor: 'rgb(200, 0, 200)',
   progressColor: 'rgb(100, 0, 100)',
   url: '/examples/audio/demo.wav',
@@ -24,3 +24,12 @@ ws.registerPlugin(
 ws.once('interaction', () => {
   ws.play()
 })
+
+/*
+<html>
+  <div id="waveform"></div>
+  <p>
+    📖 <a href="https://wavesurfer-js.org/docs/classes/plugins_spectrogram.SpectrogramPlugin">Spectrogram plugin docs</a>
+  </p>
+</html>
+*/
