+++ examples/timeline.js
@@ -5,7 +5,7 @@ import TimelinePlugin from 'https://unpkg.com/wavesurfer.js@beta/dist/plugins/ti
 
 // Create an instance of WaveSurfer
 const ws = WaveSurfer.create({
-  container: document.body,
+  container: '#waveform',
   waveColor: 'rgb(200, 0, 200)',
   progressColor: 'rgb(100, 0, 100)',
   url: '/examples/audio/audio.wav',
@@ -24,3 +24,12 @@ ws.on('interaction', () => {
 ws.on('finish', () => {
   ws.setTime(0)
 })
+
+/*
+<html>
+  <div id="waveform"></div>
+  <p>
+    📖 <a href="https://wavesurfer-js.org/docs/classes/plugins_timeline.TimelinePlugin">Timeline plugin docs</a>
+  </p>
+</html>
+*/
